Extracted BasicTypeSyntax and ListSyntax to separate their implementations
